Include configuration property so timeouts can be disabled entirely
